+++ rio/src/RIO/Prelude/Logger.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE NoImplicitPrelude #-}
@@ -20,6 +23,9 @@ module RIO.Prelude.Logger
   , setLogUseColor
   , setLogUseLoc
   , setLogFormat
+  , setLogLevelColors
+  , setLogSecondaryColor
+  , setLogAccentColors
     -- ** Standard logging functions
   , logDebug
   , logInfo
@@ -31,6 +37,8 @@ module RIO.Prelude.Logger
   , logSticky
   , logStickyDone
     -- *** With source
+    --
+    -- $withSource
   , logDebugS
   , logInfoS
   , logWarnS
@@ -50,6 +58,20 @@ module RIO.Prelude.Logger
   , noLogging
     -- ** Accessors
   , logFuncUseColorL
+  , logFuncLogLevelColorsL
+  , logFuncSecondaryColorL
+  , logFuncAccentColorsL
+    -- * Type-generic logger
+    -- $type-generic-intro
+  , glog
+  , GLogFunc
+  , gLogFuncClassic
+  , mkGLogFunc
+  , contramapMaybeGLogFunc
+  , contramapGLogFunc
+  , HasGLogFunc(..)
+  , HasLogLevel(..)
+  , HasLogSource(..)
   ) where
 
 import RIO.Prelude.Reexports hiding ((<>))
@@ -65,13 +87,17 @@ import qualified Data.Text.IO as TIO
 import Data.Bits
 import Data.ByteString.Builder (toLazyByteString, char7, byteString, hPutBuilder)
 import Data.ByteString.Builder.Extra (flush)
-import           GHC.IO.Handle.Internals         (wantWritableHandle)
 import           GHC.IO.Encoding.Types           (textEncodingName)
-import           GHC.IO.Handle.Types             (Handle__ (..))
 import qualified Data.ByteString as B
-import           System.IO                  (localeEncoding)
+import           System.IO                  (hGetEncoding, localeEncoding)
+import           System.IO.Error            (illegalOperationErrorType,
+                                             mkIOError, ioError)
 import           GHC.Foreign                (peekCString, withCString)
 import Data.Semigroup (Semigroup (..))
+
+#if MIN_VERSION_base(4,12,0)
+import Data.Functor.Contravariant
+#endif
 
 -- | The log level of a message.
 --
@@ -185,6 +211,17 @@ logOther
   -> m ()
 logOther = logGeneric "" . LevelOther
 
+-- $withSource
+--
+-- There is a set of logging functions that take an extra 'LogSource'
+-- argument to provide context, typically detailing what part of an
+-- application the message comes from.
+--
+-- For example, in verbose mode, @infoLogS "database" "connected"@ will
+-- result in
+--
+-- > [info] (database) connected
+
 -- | Log a debug level message with the given source.
 --
 -- @since 0.0.0.0
@@ -264,9 +301,17 @@ logStickyDone = logOther "sticky-done"
 -- that encompasses the logSticky->logStickyDone pairing.
 
 canUseUtf8 :: MonadIO m => Handle -> m Bool
-canUseUtf8 h = liftIO $ wantWritableHandle "canUseUtf8" h $ \h_ -> do
-  -- TODO also handle haOutputNL for CRLF
-  return $ (textEncodingName <$> haCodec h_) == Just "UTF-8"
+canUseUtf8 h = liftIO $ do
+  maybeEnc <- hGetEncoding h
+  return $ fmap textEncodingName maybeEnc == Just "UTF-8"
+
+assertWritable :: MonadIO m => Handle -> m ()
+assertWritable h = liftIO $ do
+  isOpen <- hIsOpen h -- workaround for hIsWritable bug on duplex handles
+                      -- https://gitlab.haskell.org/ghc/ghc/-/issues/26479
+  isWritable <- hIsWritable h
+  unless (isOpen && isWritable) . ioError
+    $ mkIOError illegalOperationErrorType "assertWritable" (Just h) Nothing
 
 -- | Create a 'LogOptions' value which will store its data in
 -- memory. This is primarily intended for testing purposes. This will
@@ -287,6 +332,7 @@ logOptionsMemory = do
         , logTerminal = True
         , logUseTime = False
         , logUseColor = False
+        , logColors = defaultLogColors
         , logUseLoc = False
         , logFormat = id
         , logSend = \new -> atomicModifyIORef' ref $ \old -> (old <> new, ())
@@ -296,6 +342,9 @@ logOptionsMemory = do
 -- | Create a 'LogOptions' value from the given 'Handle' and whether
 -- to perform verbose logging or not. Individiual settings can be
 -- overridden using appropriate @set@ functions.
+-- Logging output is guaranteed to be non-interleaved only for a
+-- UTF-8 'Handle' in a multi-thread environment.
+-- If the handle is not writable, an exception is thrown immediately.
 --
 -- When Verbose Flag is @True@, the following happens:
 --
@@ -312,6 +361,7 @@ logOptionsHandle
   -> Bool -- ^ Verbose Flag
   -> m LogOptions
 logOptionsHandle handle' verbose = liftIO $ do
+  assertWritable handle' -- throw IOError now, if it would already throw later
   terminal <- hIsTerminalDevice handle'
   useUtf8 <- canUseUtf8 handle'
   unicode <- if useUtf8 then return True else getCanUseUnicode
@@ -325,6 +375,7 @@ logOptionsHandle handle' verbose = liftIO $ do
 #else
     , logUseColor = verbose && terminal
 #endif
+    , logColors = defaultLogColors
     , logUseLoc = verbose
     , logFormat = id
     , logSend = \builder ->
@@ -427,10 +478,44 @@ data LogOptions = LogOptions
   , logTerminal :: !Bool
   , logUseTime :: !Bool
   , logUseColor :: !Bool
+  , logColors :: !LogColors
   , logUseLoc :: !Bool
   , logFormat :: !(Utf8Builder -> Utf8Builder)
   , logSend :: !(Builder -> IO ())
   }
+
+-- | ANSI color codes for use in the configuration of the creation of a
+-- 'LogFunc'.
+--
+-- @since 0.1.18.0
+data LogColors = LogColors
+  { -- | The color associated with each 'LogLevel'.
+    logColorLogLevels :: !(LogLevel -> Utf8Builder)
+    -- | The color of secondary content.
+  , logColorSecondary :: !Utf8Builder
+    -- | The color of accents, which are indexed by 'Int'.
+  , logColorAccents :: !(Int -> Utf8Builder)
+  }
+
+defaultLogColors :: LogColors
+defaultLogColors = LogColors
+  { logColorLogLevels = defaultLogLevelColors
+  , logColorSecondary = defaultLogSecondaryColor
+  , logColorAccents = defaultLogAccentColors
+  }
+
+defaultLogLevelColors :: LogLevel -> Utf8Builder
+defaultLogLevelColors LevelDebug = "\ESC[32m" -- Green
+defaultLogLevelColors LevelInfo = "\ESC[34m" -- Blue
+defaultLogLevelColors LevelWarn = "\ESC[33m" -- Yellow
+defaultLogLevelColors LevelError = "\ESC[31m" -- Red
+defaultLogLevelColors (LevelOther _) = "\ESC[35m" -- Magenta
+
+defaultLogSecondaryColor :: Utf8Builder
+defaultLogSecondaryColor = "\ESC[90m"  -- Bright black (gray)
+
+defaultLogAccentColors :: Int -> Utf8Builder
+defaultLogAccentColors = const "\ESC[92m" -- Bright green
 
 -- | Set the minimum log level. Messages below this level will not be
 -- printed.
@@ -468,7 +553,7 @@ setLogVerboseFormatIO :: IO Bool -> LogOptions -> LogOptions
 setLogVerboseFormatIO getVerboseLevel options =
   options { logVerboseFormat = getVerboseLevel }
 
--- | Do we treat output as a terminal. If @True@, we will enabled
+-- | Do we treat output as a terminal. If @True@, we will enable
 -- sticky logging functionality.
 --
 -- Default: checks if the @Handle@ provided to 'logOptionsHandle' is a
@@ -494,6 +579,44 @@ setLogUseTime t options = options { logUseTime = t }
 setLogUseColor :: Bool -> LogOptions -> LogOptions
 setLogUseColor c options = options { logUseColor = c }
 
+-- | ANSI color codes for 'LogLevel' in the log output.
+--
+-- Default: 'LevelDebug'   = \"\\ESC[32m\" -- Green
+--          'LevelInfo'    = \"\\ESC[34m\" -- Blue
+--          'LevelWarn'    = \"\\ESC[33m\" -- Yellow
+--          'LevelError'   = \"\\ESC[31m\" -- Red
+--          'LevelOther' _ = \"\\ESC[35m\" -- Magenta
+--
+-- @since 0.1.18.0
+setLogLevelColors :: (LogLevel -> Utf8Builder) -> LogOptions -> LogOptions
+setLogLevelColors logLevelColors options =
+  let lc = (logColors options){ logColorLogLevels = logLevelColors }
+  in  options { logColors = lc }
+
+-- | ANSI color codes for secondary content in the log output.
+--
+-- Default: \"\\ESC[90m\" -- Bright black (gray)
+--
+-- @since 0.1.18.0
+setLogSecondaryColor :: Utf8Builder -> LogOptions -> LogOptions
+setLogSecondaryColor c options =
+  let lc = (logColors options){ logColorSecondary = c }
+  in  options { logColors = lc }
+
+-- | ANSI color codes for accents in the log output. Accent colors are indexed
+-- by 'Int'.
+--
+-- Default: 'const' \"\\ESC[92m\" -- Bright green, for all indicies
+--
+-- @since 0.1.18.0
+setLogAccentColors
+  :: (Int -> Utf8Builder)  -- ^ This should be a total function.
+  -> LogOptions
+  -> LogOptions
+setLogAccentColors accentColors options =
+  let lc = (logColors options){ logColorAccents = accentColors }
+  in  options { logColors = lc }
+
 -- | Use code location in the log output.
 --
 -- Default: `True` if in verbose mode, `False` otherwise.
@@ -506,12 +629,12 @@ setLogUseLoc l options = options { logUseLoc = l }
 --
 -- Default: `id`
 --
--- @since 0.1.12.0
+-- @since 0.1.13.0
 setLogFormat :: (Utf8Builder -> Utf8Builder) -> LogOptions -> LogOptions
 setLogFormat f options = options { logFormat = f }
 
 simpleLogFunc :: LogOptions -> CallStack -> LogSource -> LogLevel -> Utf8Builder -> IO ()
-simpleLogFunc lo cs _src level msg = do
+simpleLogFunc lo cs src level msg = do
     logLevel   <- logMinLevel lo
     logVerbose <- logVerboseFormat lo
 
@@ -521,18 +644,17 @@ simpleLogFunc lo cs _src level msg = do
         timestamp <>
         getLevel logVerbose <>
         ansi reset <>
+        getSource <>
         logFormat lo msg <>
         getLoc <>
         ansi reset <>
         "\n"
   where
    reset = "\ESC[0m"
-   setBlack = "\ESC[90m"
-   setGreen = "\ESC[32m"
-   setBlue = "\ESC[34m"
-   setYellow = "\ESC[33m"
-   setRed = "\ESC[31m"
-   setMagenta = "\ESC[35m"
+   lc = logColors lo
+   levelColor = logColorLogLevels lc level
+   timestampColor = logColorSecondary lc
+   locColor = logColorSecondary lc
 
    ansi :: Utf8Builder -> Utf8Builder
    ansi xs | logUseColor lo = xs
@@ -542,7 +664,7 @@ simpleLogFunc lo cs _src level msg = do
    getTimestamp logVerbose
      | logVerbose && logUseTime lo =
        do now <- getZonedTime
-          return $ ansi setBlack <> fromString (formatTime' now) <> ": "
+          return $ ansi timestampColor <> fromString (formatTime' now) <> ": "
      | otherwise = return mempty
      where
        formatTime' =
@@ -550,22 +672,26 @@ simpleLogFunc lo cs _src level msg = do
 
    getLevel :: Bool -> Utf8Builder
    getLevel logVerbose
-     | logVerbose =
+     | logVerbose = ansi levelColor <>
          case level of
-           LevelDebug -> ansi setGreen <> "[debug] "
-           LevelInfo -> ansi setBlue <> "[info] "
-           LevelWarn -> ansi setYellow <> "[warn] "
-           LevelError -> ansi setRed <> "[error] "
+           LevelDebug -> "[debug] "
+           LevelInfo -> "[info] "
+           LevelWarn -> "[warn] "
+           LevelError -> "[error] "
            LevelOther name ->
-             ansi setMagenta <>
              "[" <>
              display name <>
              "] "
      | otherwise = mempty
 
+   getSource :: Utf8Builder
+   getSource = case src of
+     "" -> ""
+     _  -> "(" <> display src <> ") "
+
    getLoc :: Utf8Builder
    getLoc
-     | logUseLoc lo = ansi setBlack <> "\n@(" <> displayCallStack cs <> ")"
+     | logUseLoc lo = ansi locColor <> "\n@(" <> displayCallStack cs <> ")"
      | otherwise = mempty
 
 -- | Convert a 'CallStack' value into a 'Utf8Builder' indicating
@@ -651,6 +777,42 @@ utf8CharacterCount = go 0
 logFuncUseColorL :: HasLogFunc env => SimpleGetter env Bool
 logFuncUseColorL = logFuncL.to (maybe False logUseColor . lfOptions)
 
+-- | What color is the log func configured to use for each 'LogLevel'?
+--
+-- Intended for use by code which wants to optionally add additional color to
+-- its log messages.
+--
+-- @since 0.1.18.0
+logFuncLogLevelColorsL :: HasLogFunc env
+                       => SimpleGetter env (LogLevel -> Utf8Builder)
+logFuncLogLevelColorsL = logFuncL.to
+                           (maybe defaultLogLevelColors
+                                  (logColorLogLevels . logColors) . lfOptions)
+
+-- | What color is the log func configured to use for secondary content?
+--
+-- Intended for use by code which wants to optionally add additional color to
+-- its log messages.
+--
+-- @since 0.1.18.0
+logFuncSecondaryColorL :: HasLogFunc env
+                       => SimpleGetter env Utf8Builder
+logFuncSecondaryColorL = logFuncL.to
+                           (maybe defaultLogSecondaryColor
+                                  (logColorSecondary . logColors) . lfOptions)
+
+-- | What accent colors, indexed by 'Int', is the log func configured to use?
+--
+-- Intended for use by code which wants to optionally add additional color to
+-- its log messages.
+--
+-- @since 0.1.18.0
+logFuncAccentColorsL :: HasLogFunc env
+                       => SimpleGetter env (Int -> Utf8Builder)
+logFuncAccentColorsL = logFuncL.to
+                           (maybe defaultLogAccentColors
+                                  (logColorAccents . logColors) . lfOptions)
+
 -- | Disable logging capabilities in a given sub-routine
 --
 -- Intended to skip logging in general purpose implementations, where secrets
@@ -659,3 +821,190 @@ logFuncUseColorL = logFuncL.to (maybe False logUseColor . lfOptions)
 -- @since 0.1.5.0
 noLogging :: (HasLogFunc env, MonadReader env m) => m a -> m a
 noLogging = local (set logFuncL mempty)
+
+--------------------------------------------------------------------------------
+--
+-- $type-generic-intro
+--
+-- When logging takes on a more semantic meaning and the logs need to
+-- be digested, acted upon, translated or serialized upstream (to
+-- e.g. a JSON logging server), we have 'GLogFunc' (as in "generic log
+-- function"), and is accessed via 'HasGLogFunc'.
+--
+-- There is only one function to log in this system: the 'glog'
+-- function, which can log any message. You determine the log levels
+-- or severity of messages when needed.
+--
+-- Using 'RIO.Prelude.mapRIO' and 'contramapGLogFunc' (or
+-- 'contramapMaybeGLogFunc'), you can build hierarchies of loggers.
+--
+-- Example:
+--
+-- @
+-- import RIO
+--
+-- data DatabaseMsg = Connected String | Query String | Disconnected deriving Show
+-- data WebMsg = Request String | Error String | DatabaseMsg DatabaseMsg deriving Show
+-- data AppMsg = InitMsg String | WebMsg WebMsg deriving Show
+--
+-- main :: IO ()
+-- main =
+--   runRIO
+--     (mkGLogFunc (\stack msg -> print msg))
+--     (do glog (InitMsg "Ready to go!")
+--         runWeb
+--           (do glog (Request "/foo")
+--               runDB (do glog (Connected "127.0.0.1")
+--                         glog (Query "SELECT 1"))
+--               glog (Error "Oh noes!")))
+--
+-- runDB :: RIO (GLogFunc DatabaseMsg) () -> RIO (GLogFunc WebMsg) ()
+-- runDB = mapRIO (contramapGLogFunc DatabaseMsg)
+--
+-- runWeb :: RIO (GLogFunc WebMsg) () -> RIO (GLogFunc AppMsg) ()
+-- runWeb = mapRIO (contramapGLogFunc WebMsg)
+-- @
+--
+-- If we instead decided that we only wanted to log database queries,
+-- and not bother the upstream with connect/disconnect messages, we
+-- could simplify the constructor to @DatabaseQuery String@:
+--
+-- @
+-- data WebMsg = Request String | Error String | DatabaseQuery String deriving Show
+-- @
+--
+-- And then @runDB@ could use 'contramapMaybeGLogFunc' to parse only queries:
+--
+-- @
+-- runDB =
+--   mapRIO
+--     (contramapMaybeGLogFunc
+--        (\msg ->
+--           case msg of
+--             Query string -> pure (DatabaseQuery string)
+--             _ -> Nothing))
+-- @
+--
+-- This way, upstream only has to care about queries and not
+-- connect/disconnect constructors.
+
+-- | An app is capable of generic logging if it implements this.
+--
+-- @since 0.1.13.0
+class HasGLogFunc env where
+  type GMsg env
+  gLogFuncL :: Lens' env (GLogFunc (GMsg env))
+
+-- | Quick way to run a RIO that only has a logger in its environment.
+--
+-- @since 0.1.13.0
+instance HasGLogFunc (GLogFunc msg) where
+  type GMsg (GLogFunc msg) = msg
+  gLogFuncL = id
+
+-- | A generic logger of some type @msg@.
+--
+-- Your 'GLocFunc' can re-use the existing classical logging framework
+-- of RIO, and/or implement additional transforms,
+-- filters. Alternatively, you may log to a JSON source in a database,
+-- or anywhere else as needed. You can decide how to log levels or
+-- severities based on the constructors in your type. You will
+-- normally determine this in your main app entry point.
+--
+-- @since 0.1.13.0
+newtype GLogFunc msg = GLogFunc (CallStack -> msg -> IO ())
+
+#if MIN_VERSION_base(4,12,0)
+-- https://hackage.haskell.org/package/base-4.12.0.0/docs/Data-Functor-Contravariant.html
+
+-- | Use this instance to wrap sub-loggers via 'RIO.mapRIO'.
+--
+-- The 'Contravariant' class is available in base 4.12.0.
+--
+-- @since 0.1.13.0
+instance Contravariant GLogFunc where
+  contramap = contramapGLogFunc
+  {-# INLINABLE contramap #-}
+#endif
+
+-- | Perform both sets of actions per log entry.
+--
+-- @since 0.1.13.0
+instance Semigroup (GLogFunc msg) where
+  GLogFunc f <> GLogFunc g = GLogFunc (\a b -> f a b *> g a b)
+
+-- | 'mempty' peforms no logging.
+--
+-- @since 0.1.13.0
+instance Monoid (GLogFunc msg) where
+  mempty = mkGLogFunc $ \_ _ -> return ()
+  mappend = (<>)
+
+-- | A vesion of 'contramapMaybeGLogFunc' which supports filering.
+--
+-- @since 0.1.13.0
+contramapMaybeGLogFunc :: (a -> Maybe b) -> GLogFunc b -> GLogFunc a
+contramapMaybeGLogFunc f (GLogFunc io) =
+  GLogFunc (\stack msg -> maybe (pure ()) (io stack) (f msg))
+{-# INLINABLE contramapMaybeGLogFunc #-}
+
+-- | A contramap. Use this to wrap sub-loggers via 'RIO.mapRIO'.
+--
+-- If you are on base > 4.12.0, you can just use 'contramap'.
+--
+-- @since 0.1.13.0
+contramapGLogFunc :: (a -> b) -> GLogFunc b -> GLogFunc a
+contramapGLogFunc f (GLogFunc io) = GLogFunc (\stack msg -> io stack (f msg))
+{-# INLINABLE contramapGLogFunc #-}
+
+-- | Make a custom generic logger. With this you could, for example,
+-- write to a database or a log digestion service. For example:
+--
+-- > mkGLogFunc (\stack msg -> send (Data.Aeson.encode (JsonLog stack msg)))
+--
+-- @since 0.1.13.0
+mkGLogFunc :: (CallStack -> msg -> IO ()) -> GLogFunc msg
+mkGLogFunc = GLogFunc
+
+-- | Log a value generically.
+--
+-- @since 0.1.13.0
+glog ::
+     (MonadIO m, HasCallStack, HasGLogFunc env, MonadReader env m)
+  => GMsg env
+  -> m ()
+glog t = do
+  GLogFunc gLogFunc <- view gLogFuncL
+  liftIO (gLogFunc callStack t)
+{-# INLINABLE glog #-}
+
+--------------------------------------------------------------------------------
+-- Integration with classical logger framework
+
+-- | Level, if any, of your logs. If unknown, use 'LogOther'. Use for
+-- your generic log data types that want to sit inside the classic log
+-- framework.
+--
+-- @since 0.1.13.0
+class HasLogLevel msg where
+  getLogLevel :: msg -> LogLevel
+
+-- | Source of a log. This can be whatever you want. Use for your
+-- generic log data types that want to sit inside the classic log
+-- framework.
+--
+-- @since 0.1.13.0
+class HasLogSource msg where
+  getLogSource :: msg -> LogSource
+
+-- | Make a 'GLogFunc' via classic 'LogFunc'. Use this if you'd like
+-- to log your generic data type via the classic RIO terminal logger.
+--
+-- @since 0.1.13.0
+gLogFuncClassic ::
+     (HasLogLevel msg, HasLogSource msg, Display msg) => LogFunc -> GLogFunc msg
+gLogFuncClassic (LogFunc {unLogFunc = io}) =
+  mkGLogFunc
+    (\theCallStack msg ->
+       liftIO
+         (io theCallStack (getLogSource msg) (getLogLevel msg) (display msg)))