+++ rio/src/RIO/Prelude/Types.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE CPP                #-}
+#if MIN_VERSION_base(4,17,0)
+{-# LANGUAGE ExplicitNamespaces #-}
+#endif
+
 module RIO.Prelude.Types
   (
     -- * @base@
@@ -133,9 +138,15 @@ module RIO.Prelude.Types
     -- **** @Functor@
     -- | Re-exported from "Data.Functor":
   , Data.Functor.Functor
+    -- **** @Bifunctor@
+    -- | Re-exported from "Data.Bifunctor":
+  , Data.Bifunctor.Bifunctor
     -- **** @Foldable@
     -- | Re-exported from "Data.Foldable":
   , Data.Foldable.Foldable
+    -- **** @Bifoldable@
+    -- | Re-exported from "Data.Bifoldable":
+  , Data.Bifoldable.Bifoldable
     -- **** @Semigroup@
     -- | Re-exported from "Data.Semigroup":
   , Data.Semigroup.Semigroup
@@ -151,6 +162,9 @@ module RIO.Prelude.Types
     -- **** @Traversable@
     -- | Re-exported from "Data.Traversable":
   , Data.Traversable.Traversable
+    -- **** @Bitraversable@
+    -- | Re-exported from "Data.Bitraversable":
+  , Data.Bitraversable.Bitraversable
     -- **** @Monad@
     -- | Re-exported from "Control.Monad":
   , Control.Monad.Monad
@@ -189,6 +203,11 @@ module RIO.Prelude.Types
     -- | Re-exported from "GHC.Stack":
   , GHC.Stack.HasCallStack
 
+#if MIN_VERSION_base(4,17,0)
+    -- *** The equality types
+    -- | Re-exported from "Data.Type.Equality":
+  , type (Data.Type.Equality.~)
+#endif
 
     -- * @deepseq@
     -- ** @NFData@
@@ -287,17 +306,6 @@ module RIO.Prelude.Types
   , Control.Monad.Primitive.PrimMonad (PrimState)
   ) where
 
-import qualified Control.Monad.Primitive (PrimMonad(..))
-import qualified Data.ByteString (ByteString)
-import qualified Data.ByteString.Builder (Builder)
-import qualified Data.Monoid (Monoid)
-import qualified Data.Semigroup (Semigroup)
-import qualified Data.String (IsString, String)
-import qualified Data.Text (Text)
-import qualified Data.Typeable
-import qualified System.IO
-
-import qualified Data.Vector.Unboxed (Unbox)
 import qualified RIO.Prelude.Renames
 
 import qualified Control.Applicative
@@ -308,9 +316,15 @@ import qualified Control.Exception.Base
 import qualified Control.Monad
 import qualified Control.Monad.Catch
 import qualified Control.Monad.Fail
+import qualified Control.Monad.Primitive (PrimMonad(..))
 import qualified Control.Monad.Reader
 import qualified Control.Monad.ST
+import qualified Data.Bifoldable
+import qualified Data.Bifunctor
+import qualified Data.Bitraversable
 import qualified Data.Bool
+import qualified Data.ByteString (ByteString)
+import qualified Data.ByteString.Builder (Builder)
 import qualified Data.ByteString.Short
 import qualified Data.Char
 import qualified Data.Data
@@ -331,14 +345,25 @@ import qualified Data.List
 import qualified Data.List.NonEmpty
 import qualified Data.Map.Strict
 import qualified Data.Maybe
+import qualified Data.Monoid (Monoid)
 import qualified Data.Ord
 import qualified Data.Proxy
 import qualified Data.Ratio
+import qualified Data.Semigroup (Semigroup)
 import qualified Data.Sequence
 import qualified Data.Set
+import qualified Data.String (IsString, String)
+import qualified Data.Text (Text)
 import qualified Data.Text.Encoding.Error
 import qualified Data.Traversable
+-- See https://errors.haskell.org/messages/GHC-12003/
+-- From GHC 9.4.1, the type equality operator ~ is a regular type operator.
+#if MIN_VERSION_base(4,17,0)
+import qualified Data.Type.Equality (type (~))
+#endif
+import qualified Data.Typeable
 import qualified Data.Vector
+import qualified Data.Vector.Unboxed (Unbox)
 import qualified Data.Void
 import qualified Data.Word
 import qualified Foreign.Storable
@@ -347,8 +372,9 @@ import qualified GHC.Stack
 import qualified Numeric.Natural
 import qualified Prelude
 import qualified System.Exit
+import qualified System.IO
 import qualified Text.Read
 import qualified Text.Show
 
 -- Bring instances for some of the unliftio types in scope, so they can be documented here.
-import UnliftIO ()
+import           UnliftIO ()