+++ rio-orphans/ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog for rio-orphans
 
+## 0.1.2.0
+
+* Add an instance for `MonadLoggerIO` typeclass, from the `monad-logger` library
+
 ## 0.1.1.0
 
 * Add an instance for `MonadLogger` typeclass, from the `monad-logger` library